fix the recursive bug
