+++ lib/octophus.js
@@ -6,8 +6,9 @@ import { CompositeDisposable } from 'atom';
 import packageConfig from './config.json';
 import githubAPI from './github-integrate/auth';
 import Repository from './github-integrate/repository';
+import Directory from './directory';
+import TreeView from './view/tree-view';
 
-var TreeView, Directory;
 
 export default {
   config: packageConfig,
@@ -16,11 +17,6 @@ export default {
   treeView: null,
 
   activate(state) {
-    if (!TreeView) {
-      TreeView = require('./view/tree-view');
-      Directory = require('./directory');
-    }
-
     atom.project.octophusdir = new Directory({
       parent: null,
       name: "/"
@@ -63,15 +59,13 @@ export default {
           let repo = fullrepo.split('/')[1].split(':')[0];
           let brench = fullrepo.split(':')[1];
           let repository = new Repository(user, repo, brench);
-          repository.getDirectory(function (ghDir) {
-            name = ghDir.repo.name;
-            dir = new Directory({
-              parent: null,
-              name: name,
-              ghDir: ghDir
-            });
-            dir.open(false, function() {atom.project.octophustree.setRoot(dir);} );
-          });
+          repository.getDirectory()
+            .then(ghDir => new Directory({
+                parent: null,
+                name: ghDir.repo.name,
+                ghDir: ghDir
+              }).open())
+            .then(() => atom.project.octophustree.setRoot(dir));
         });
         dialog.attach();
       }
